+++ apps/api/src/app/portfolio/portfolio.service.ts
@@ -1873,6 +1873,7 @@ export class PortfolioService {
 
     const {
       currentValueInBaseCurrency,
+      dividendYieldTrailingTwelveMonths,
       totalInvestment,
       totalInvestmentWithCurrencyEffect
     } = await portfolioCalculator.getSnapshot();
@@ -1977,6 +1978,7 @@ export class PortfolioService {
       annualizedPerformancePercentWithCurrencyEffect,
       cash,
       dateOfFirstActivity,
+      dividendYieldTrailingTwelveMonths,
       excludedAccountsAndActivities,
       netPerformance,
       netPerformancePercentage,